Updated: Added the onClickListener for the titleObject's Custom Listview Image Button Adapter thing
POSSIBLE FORSEEABLE PROBLEMS:
create new title object may need altering as there will originally be no titleString...
POSSIBLE SOLUTION:
create function that assigns only id and date created but blank string for title. Take first toDoObject as title if no title is input.

TODO:
Create an onclick listener for the delete button function in the titlePage Adapter class
Create a button to save the objects on todoPage (save and return or something)
Think of a way to allow a user to edit the object real-time (would using a onTextChanged Listener be inefficient? -- Maybe but would work)